+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+### Fixed
+- Use deserialize cache stub when creating a `Jsona` instance
+
 ## [1.3.0] - 2019-03-01
 
 [Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.3.0...HEAD